Rename `common` to `flow_sdk`, update other connectors to use this, fix logging
