+++ source-airtable/pyproject.toml
@@ -9,8 +9,8 @@ airbyte-cdk = "0.51.14"
 mypy = "^1.5"
 python = "^3.11"
 debugpy = "^1.8.0"
-source_airtable = { path = "source_airtable" }
-common = {path="../python"}
+source_airtable = { path = "source_airtable", develop=true }
+flow-sdk = {path="../python", develop=true}
 
 [build-system]
 requires = ["poetry-core"]
