+++ source-airtable/test.flow.yaml
@@ -1,43 +1,20 @@
 ---
 captures:
   acmeCo/source-airtable:
-    shards:
-      logLevel: debug
     endpoint:
-      connector:
-        image: ghcr.io/estuary/source-airtable:local
-        config:
-          pokemon_name: pikachu
-      # local:
-      #   command:
-      #     - python
-      #     # - "-m"
-      #     # - "debugpy"
-      #     # - "--listen"
-      #     # - "0.0.0.0:5678"
-      #     # - "--wait-for-client"
-      #     - "-m"
-      #     - source-airtable
+      # connector:
+      #   image: ghcr.io/estuary/source-airtable:local
       #   config:
-      #     pokemon_name: pikachu
-
-    bindings:
-      - resource:
-          stream: airtable
-          syncMode: full_refresh
-        target: acmeCo/airtable
-collections:
-  acmeCo/airtable:
-    schema:
-      type: object
-      x-infer-schema: true
-      properties:
-        _meta:
-          type: object
-          properties:
-            row_id:
-              type: integer
-          required:
-            - row_id
-    key:
-      - /_meta/row_id
+      #       credentials:
+      #         auth_method: api_key
+      #         api_key: pat... 
+      local:
+        command:
+          - python
+          - "-m"
+          - source-airtable
+        config:
+          credentials:
+            auth_method: api_key
+            api_key: pat...
+    bindings: []
