+++ source-yahoo-finance/pyproject.toml
@@ -12,6 +12,7 @@ orjson = "^3.9.7"
 python = ">=3.11,<3.12"
 tap-investing = {git = "https://gitlab.com/DouweM/tap-investing.git", rev = "1a12a4d"}
 debugpy = "^1.8.0"
+flow-sdk = {path="../python", develop=true}
 
 [build-system]
 requires = ["poetry-core"]
